Vivek/jwt auth

+++ server/controllers/ourDBController.js
@@ -4,16 +4,19 @@ import moment from 'moment';
 const ourDBController = {};
 
 ourDBController.queryPush = async (req, res, next) => {
-  const { _id, querystring, queryname, querydelay, querycount, querymetrics, averagetime } = res.locals.metrics;
-  const stringquerymetrics = JSON.stringify(querymetrics);
-  const date = moment().format();
-  const string = `INSERT INTO metrics (querystring, querymetrics, queryname, querycount, querydelay, averagetime, users_id, created_at)
-  VALUES ($1, $2, $3, $4, $5, $6, $7, $8);`;
+  const { _id, queryString, queryName, queryDelay, queryCount, queryMetrics, averageTime } = res.locals.metrics;
+  const stringQueryMetrics = JSON.stringify(queryMetrics);
+  const createdAt = moment().format();
+  const string = { text: `INSERT INTO metrics (queryString, queryMetrics, queryName, queryCount, queryDelay, averageTime, users_id, createdAt)
+  VALUES ($1, $2, $3, $4, $5, $6, $7, $8);`, values: [queryString, stringQueryMetrics, queryName, queryCount, queryDelay, averageTime, _id, createdAt] };
   try {
-    const result = await ourDBModel(string, [querystring, stringquerymetrics, queryname, querycount, querydelay, averagetime, _id, date]);
-    console.log(result);
-    res.locals.querymetrics = result;
-    res.locals.metrics.created_at = date;
+    //up to scrap lines 14-15
+    const result = await ourDBModel(string);
+
+    res.locals.queryMetrics = result;
+    
+    res.locals.metrics.createdAt = createdAt;
+
     return next();
   } catch (err) {
     return next({
@@ -24,15 +27,17 @@ ourDBController.queryPush = async (req, res, next) => {
   }
 };
 
+// get query data for user
 ourDBController.queryGet = async (req, res, next) => {
-  //We might need to add queryname to this body. 
-  const { _id } = req.body;
+  const { _id } = req.user;
+  
   const string = {text: 'SELECT * FROM metrics WHERE users_id = $1', values:[_id] };
   try {
     const result = await ourDBModel(string);
     const resultData = result.rows;
     const returnDataMetrics = resultData.map(({ users_id, ...rest }) => rest);  
-    res.locals.getmetrics = returnDataMetrics;
+    res.locals.getMetrics = returnDataMetrics;
+    
     return next();
   } catch (err) {
     return next({
@@ -45,10 +50,13 @@ ourDBController.queryGet = async (req, res, next) => {
 
 ourDBController.deleteQueryById = async (req, res, next) => {
   const { _id } = req.body;
-  const string = {text: 'DELETE FROM metrics WHERE _id = $1', values:[_id] };
+ 
+  const userId = req.user._id;
+  
+  const string = {text: 'DELETE FROM metrics WHERE _id = $1 AND users_id = $2', values:[_id,userId] };
   try {
     const result = await ourDBModel(string);
-    console.log(result);  
+     
     return next();
   } catch (err) {
     return next({
@@ -60,11 +68,10 @@ ourDBController.deleteQueryById = async (req, res, next) => {
 };
 
 ourDBController.deleteQueryByName = async (req, res, next) => {
-  const { queryname } = req.body;
-  const string = {text: 'DELETE FROM metrics WHERE queryname = $1', values:[queryname] };
+  const { queryName } = req.body;
+  const string = {text: 'DELETE FROM metrics WHERE queryName = $1', values:[queryName] };
   try {
     const result = await ourDBModel(string);
-    console.log(result);  
     return next();
   } catch (err) {
     return next({
